* Added: gitignore * Changed: folder structure * Fixed: "fris" menu is now working
